backend_neu/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
numpy==2.2.6
pandas==2.2.3
psycopg[binary]==3.2.9
python-dotenv==1.2.2
